requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-sequences==2.5
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0